backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
httpx==0.27.0
pydantic-settings==2.2.1
loguru==0.7.2
tenacity==8.2.3
python-dotenv==1.0.1
